+++ src/main/java/net/geforcemods/securitycraft/util/InventoryUtils.java
@@ -9,8 +9,8 @@ import net.geforcemods.securitycraft.blocks.InventoryScannerFieldBlock;
 import net.minecraft.core.NonNullList;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.world.Container;
-import net.minecraft.world.entity.EntityEquipment;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ItemStackTemplate;
 import net.minecraft.world.item.component.BundleContents;
@@ -197,21 +197,29 @@ public class InventoryUtils {
 			};
 		}
 
-		static ItemAccess forEntityEquipment(EntityEquipment equipment, EquipmentSlot... validSlots) {
+		static ItemAccess forEntityEquipment(LivingEntity entity) {
 			return new ItemAccess() {
 				@Override
 				public int size() {
-					return validSlots.length;
+					return EquipmentSlot.values().length;
 				}
 
 				@Override
 				public ItemStack getItem(int slot) {
-					return equipment.get(validSlots[slot]);
+					int size = size();
+
+					if (slot < 0 || slot >= size)
+						return ItemStack.EMPTY;
+
+					return entity.getItemBySlot(EquipmentSlot.values()[slot]);
 				}
 
 				@Override
 				public void set(int slot, ItemStack stack) {
-					equipment.set(validSlots[slot], stack);
+					int size = size();
+
+					if (slot >= 0 || slot < size)
+						entity.setItemSlot(EquipmentSlot.values()[slot], stack);
 				}
 			};
 		}